+++ dsu/src/main/java/tp1/fpaa/dsu/DSUUnionRank.java
@@ -23,7 +23,7 @@ public class DSUUnionRank implements DSU {
      * rank[x] é uma cota superior da altura da árvore enraizada em x. Usado para
      * balanceamento.
      */
-    private final byte[] rank;
+    private final int[] rank;
 
     /** Número máximo de elementos. Índices válidos: 0 até capacity - 1. */
     private final int capacity;
@@ -44,7 +44,7 @@ public class DSUUnionRank implements DSU {
         }
         this.capacity = n;
         this.parent = new int[n];
-        this.rank = new byte[n];
+        this.rank = new int[n];
     }
 
     @Override
@@ -82,7 +82,7 @@ public class DSUUnionRank implements DSU {
     }
 
     /** Escreve rank[x] e contabiliza o acesso se métricas estiverem ativas. */
-    private void writeRank(int x, byte value) {
+    private void writeRank(int x, int value) {
         if (metrics != null)
             metrics.incParentAccess();
         rank[x] = value;
@@ -107,7 +107,7 @@ public class DSUUnionRank implements DSU {
     @Override
     public void makeSet(int x) {
         writeParent(x, x);
-        writeRank(x, (byte) 0);
+        writeRank(x, 0);
     }
 
     /**
@@ -151,7 +151,7 @@ public class DSUUnionRank implements DSU {
         } else {
             writeParent(x, y);
             if (rx == ry) {
-                writeRank(y, (byte) (ry + 1));
+                writeRank(y, ry + 1);
             }
         }
     }
